extract data structure for orig/current scen
